# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetrainingTicketCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "human_review_confirmed": {
            "const": true,
            "title": "Human Review Confirmed",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "human_review_confirmed"
        ],
        "title": "RetrainingTicketCreateRequest",
        "type": "object"
      },
      "RetrainingTicketResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "human_review_confirmed": {
            "title": "Human Review Confirmed",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "monitoring_snapshot_id": {
            "title": "Monitoring Snapshot Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requested_by_id": {
            "title": "Requested By Id",
            "type": "string"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By Id"
          },
          "status": {
            "$ref": "#/components/schemas/RetrainingTicketStatus"
          }
        },
        "required": [
          "id",
          "monitoring_snapshot_id",
          "reason",
          "human_review_confirmed",
          "status",
          "requested_by_id",
          "reviewed_by_id",
          "review_note",
          "created_at",
          "reviewed_at"
        ],
        "title": "RetrainingTicketResponse",
        "type": "object"
      },
      "RetrainingTicketReviewDecision": {
        "enum": [
          "approve",
          "reject"
        ],
        "title": "RetrainingTicketReviewDecision",
        "type": "string"
      },
      "RetrainingTicketReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/RetrainingTicketReviewDecision"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "decision"
        ],
        "title": "RetrainingTicketReviewRequest",
        "type": "object"
      },
      "RetrainingTicketStatus": {
        "enum": [
          "open",
          "approved",
          "rejected"
        ],
        "title": "RetrainingTicketStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Human-in-the-loop loan default decision support with auditable model scores, explanations, officer feedback, and drift state. Decision support only; a qualified human makes the final decision.",
    "title": "Drift Aware Loan Default API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/retraining-tickets": {
      "get": {
        "description": "List review requests, newest first, for authorized monitoring staff.",
        "operationId": "list_retraining_tickets_api_v1_retraining_tickets_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RetrainingTicketResponse"
                  },
                  "title": "Response List Retraining Tickets Api V1 Retraining Tickets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Retraining Tickets",
        "tags": [
          "retraining"
        ]
      },
      "post": {
        "description": "Log a human-confirmed review request without enqueueing any work.",
        "operationId": "create_retraining_ticket_api_v1_retraining_tickets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrainingTicketCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrainingTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Retraining Ticket",
        "tags": [
          "retraining"
        ]
      }
    },
    "/api/v1/retraining-tickets/{ticket_id}/review": {
      "post": {
        "description": "Record an administrative disposition only; no model work is invoked.",
        "operationId": "review_retraining_ticket_api_v1_retraining_tickets__ticket_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrainingTicketReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrainingTicketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Retraining Ticket",
        "tags": [
          "retraining"
        ]
      }
    },
...
  }
}